Move all the CFI analysis implementations to lib
