+++ llvm/lib/MC/CFIAnalysis/CMakeLists.txt
@@ -1,8 +1,7 @@
 add_llvm_component_library(LLVMMCCFIAnalysis
   CFIState.cpp
-  # CFIAnalysisMCStreamer.cpp
-  # CFIAnalysis.cpp
-  # ExtendedMCInstrAnalysis.cpp
+  CFIAnalysisMCStreamer.cpp
+  CFIAnalysis.cpp
 
   LINK_COMPONENTS
   MC
